Fix build, crash-on-save, and make the core card flow usable
The app did not compile and, once building, crashed whenever a card with
faces, images, or stickers was saved. This gets it to a shippable state.

Build fixes:
- Remove duplicate ImageSource enum and duplicate PhotoLibraryManager
  overloads (PlatformImage is UIImage on iOS, so they collided)
- Rewrite FaceImportView against the real FaceDetectionService API
  (async, returns [DetectedFace]) instead of a nonexistent one
- Fix CloudKitSyncService against the current models: CKRecord cannot hold
  dictionaries, imageData/thumbnailData are non-optional, CardImage has no
  thumbnail, and SyncStatus is not Equatable
- Fix CardPreviewView to use SwiftUI Canvas APIs rather than UIKit drawing,
  so background and text actually render
- Add missing imports, fix throwing ModelContainer previews, drop two unused
  Example files

Crash fix (shipping blocker):
- FaceImage/CardImage/StickerReference stored CGPoint/CGSize directly, which
  SwiftData cannot persist ("Composite Coder only supports Keyed Container").
  Store components as Doubles and expose the geometry as computed properties.
  Same treatment for CardImage.source, matching UserPreferences' existing
  pattern. Card creation with any face/image/sticker previously crashed.
- Don't fatalError on launch when CloudKit is unavailable; fall back to local
  then in-memory storage
- Give CardImage a visible default size so added images aren't 0x0

Product:
- Home was a dead end with no way to create anything. Add "Create a Card" and
  "Make Cards from a Photo" actions plus an empty state
- Wire AI saying generation into the card editor; it was unreachable
- Add an on-device saying generator so the headline feature works with no
  backend deployed (the API URL was still the placeholder) and when offline.
  Set APIBaseURL in Info.plist to use a real backend instead

Tests:
- Repair the test target, which did not compile (wrong response type names,
  stale model APIs, missing imports, #Predicate captures)
- Replace placeholder APIService tests with real URLProtocol-backed ones
- Add regression tests for the SwiftData persistence crash, tests for the
  local generator, and a golden-path UI test for create -> generate -> save

66 unit tests and the UI tests pass; 84 test cases green overall.

+++ My Funny Valentine/Services/ImageManager.swift
@@ -160,7 +160,7 @@ actor ImageManager {
                 }
             }
         }
-        return (totalSize, maxTotalStorageBytes)
+        return (totalSize, Int64(maxTotalStorageBytes))
     }
 
     /// Clear memory cache (e.g., on memory warning)
